Fix transient condition when device path is not yet available on host
We currently fail with an internal error if the device path is not ready
on the host yet. The error orrurs when we try to mount the volume. This
leads to unnecessary error events.

In this PR we try to address this issue by return an `Unavailable` error
in case the device on the host is not ready yet. This will trigger a
re-try.